requirements.txt
python-binance
kaleido
pillow
cufflinks
python-telegram
pandas
matplotlib
plotly
mplfinance